refactor, remove consumer from OverridesDependencies class
